requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
bcrypt
email_validator
sendgrid
python-multipart
pandas
openpyxl
pytz
mailersend



